feat: implementa navegação pública, auth e privada
