+++ client/navigation/AppNavigator.js
@@ -1,17 +1,14 @@
-import React from 'react';
-import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import LoginCPFScreen from '../screens/Login/LoginCPFScreen';
-import LoginSmsScreen from '../screens/Login/LoginSmsScreen';
-import HomeScreen from '../screens/Home/HomeScreen';
-
-const Stack = createNativeStackNavigator();
+import React, { useState } from 'react';
+import AuthNavigator from './AuthNavigator';
+import PrivateNavigator from './PrivateNavigator';
+import PublicNavigator from './PublicNavigator';
 
 export default function AppNavigator() {
-  return (
-    <Stack.Navigator initialRouteName="LoginCPF" screenOptions={{headerShown:false}}>
-      <Stack.Screen name="LoginCPF" component={LoginCPFScreen} />
-      <Stack.Screen name="LoginSms" component={LoginSmsScreen} />
-      <Stack.Screen name="Home" component={HomeScreen} />
-    </Stack.Navigator>
-  );
+  const [userLogged, setUserLogged] = useState(false); // temporário
+
+  if (userLogged) {
+    return <PrivateNavigator />;
+  } else {
+    return <PublicNavigator />; 
+  }
 }
